Synchronize PCM access and transport release
This commit revises synchronization between D-Bus PCM access and the
transport release procedure in order to use PCM related mutex only. The
advantage of such implementation is a possibility to have more than one
PCM attached to a single transport object.

There is also an introduction of a transport IO thread state machine
which allows to check current state and then acts accordingly.